resume_generator/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5
httpx==0.28.1

# AI providers
anthropic==0.43.0
google-generativeai==0.8.3

# Parsing
pymupdf==1.25.2
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3

# Export
reportlab==4.2.5
